test: share extension harness setup
